+++ src/main/java/org/morphix/lang/Nullables.java
@@ -22,7 +22,9 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 
+import org.morphix.lang.function.Runnables;
 import org.morphix.lang.function.SetterFunction;
+import org.morphix.lang.function.Suppliers;
 import org.morphix.reflection.Constructors;
 
 /**
@@ -100,10 +102,8 @@ public final class Nullables {
 	 * @return the given value if the value is non-null
 	 */
 	public static <T, E extends Throwable> T nonNullOrThrow(final T value, final Supplier<E> throwableSupplier) {
-		if (null != value) {
-			return value;
-		}
-		return Unchecked.Undeclared.reThrow(throwableSupplier.get());
+		// JaCoco false positive on missed branch: the method reThrow always throws an exception and functionally is covered
+		return null != value ? value : Unchecked.Undeclared.reThrow(throwableSupplier.get());
 	}
 
 	/**
@@ -350,7 +350,7 @@ public final class Nullables {
 	 * @return a supplier which supplies a null value
 	 */
 	public static <T> Supplier<T> supplyNull() {
-		return () -> null;
+		return Suppliers.supplyNull();
 	}
 
 	/**
@@ -362,10 +362,7 @@ public final class Nullables {
 	 * @return a supplier which runs the runnable and supplies a null value
 	 */
 	public static <T> Supplier<T> supplyNull(final Runnable runnable) {
-		return () -> {
-			runnable.run();
-			return null;
-		};
+		return Runnables.compose(runnable, supplyNull());
 	}
 
 	/**
@@ -635,6 +632,7 @@ public final class Nullables {
 		 *
 		 * @param failFastErrorMessage exception message
 		 * @return value if value is not <code>null</code>
+		 * @throws IllegalStateException when value is null
 		 */
 		public T valueOrError(final String failFastErrorMessage) {
 			if (null != value) {
@@ -674,6 +672,7 @@ public final class Nullables {
 		 * @param predicate value predicate
 		 * @param failFastErrorMessage error message
 		 * @return value if value is not <code>null</code> and matches predicate
+		 * @throws IllegalStateException when value is null or does not match predicate
 		 */
 		public T valueWhenOrError(final Predicate<T> predicate, final String failFastErrorMessage) {
 			if (null != value && predicate.test(value)) {