+++ src/test/java/org/morphix/lang/JavaObjectsTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -18,6 +18,8 @@ import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
+import org.morphix.reflection.Constructors;
+import org.morphix.utils.Tests;
 
 /**
  * Test class for {@link JavaObjects}.
@@ -34,6 +36,13 @@ class JavaObjectsTest {
 
 		assertThat(bubu, notNullValue());
 		assertThat(bubu, equalTo("Bubu"));
+	}
+
+	@Test
+	void shouldThrowExceptionWhenTryingToInstantiate() {
+		UnsupportedOperationException e = Tests.verifyDefaultConstructorThrows(JavaObjects.class);
+
+		assertThat(e.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
 	}
 
 	@Test